SISTEMA_SUELDOS_PORTABLE/_internal/03_OTROS/requirements.txt
pandas
openpyxl
numpy
PyPDF2
reportlab
Pillow
num2words
pikepdf
pywin32
customtkinter
